document-service/requirements.txt
fastapi==0.104.1
uvicorn==0.23.2
pydantic==2.4.2
pydantic-settings==2.0.3
sqlalchemy==2.0.22
alembic==1.12.0
psycopg2-binary==2.9.9
python-multipart==0.0.6
python-jose==3.3.0
passlib==1.7.4
minio==7.1.17
rq==1.15.1
redis==5.0.1
httpx==0.25.0
pyclamd==0.4.0
apscheduler==3.10.4
python-dotenv==1.0.0
tenacity==8.2.3
pyyaml==6.0.1
loguru==0.7.2
shortuuid==1.0.11
python-slugify==8.0.1
